Prepare to remove previous implementation of nndescent/search/utils
