+++ src/NearestNeighborDescent.jl
@@ -6,15 +6,9 @@ using LightGraphs
 
 include("knn_graph/KNNGraphs.jl")
 using .KNNGraphs
-include("_utils.jl")
-include("_descent.jl")
-
 include("utils.jl")
-include("nn_descent.jl")
-include("brute_knn.jl")
+include("descent.jl")
 
-# main exports
-export DescentGraph
 export nndescent, search
 # KNNGraphs exports
 export KNNGraphs
