+++ src/descent.jl
@@ -111,3 +111,12 @@ function local_join!(graph::LockHeapKNNGraph, data, metric::PreMetric; sample_ra
     end
     return count[]
 end
+
+"""
+    search(graph, queries, n_neighbors; max_candidates) -> indices, distances
+
+Search the kNN `graph` for the nearest neighbors of the points in `queries`.
+`max_candidates` controls how large the candidate queue should be (min `n_neighbors`);
+larger values increase accuracy at the cost of speed.
+"""
+function search end
